requirements.txt
numpy>=1.26.2
tensorflow>=2.15.0
Pillow>=10.1.0
keras>=2.15.0
transformers~=4.12.5
pytorch~=2.1.2